Move features auto-config to autoconfigure pkg
Relocated `ConfigurationPropertyHasFeaturesAutoConfiguration` (and its related tests) from `...client.actuator` to `...client.actuator.autoconfigure` to align with auto-configuration package conventions. Updated `spring.factories` and `AutoConfiguration.imports` to reference the new fully qualified class name, and adjusted imports (including `FeaturesProperties`) after the package move.
